+++ src/components/App/OnboardingScreen/PasscodePage/PasscodePage.tsx
@@ -2,9 +2,9 @@ import styled from 'styled-components';
 import Page, {PageTransitionProps} from "../../../UI/Layout/Pages/Locked-Register/Page";
 import {FormattedMessage} from "react-intl";
 import StepButtons from '../../../UI/StepButtons/StepButtons';
-import PasscodeInput from "../../../UI/Inputs/PasscodeInput/PasscodeInput";
+import PasscodeInput, {PasscodeInputProps} from "../../../UI/Inputs/PasscodeInput/PasscodeInput";
 import Keypad, {KeypadClickHandlers} from "../../../UI/Keypad/Keypad/Keypad";
-import { ReactNode } from 'react';
+import {ReactNode} from 'react';
 
 const StyledContainer = styled.div`
   height: 100%;
@@ -14,22 +14,30 @@ const StyledContainer = styled.div`
 type PasscodeProps = {
     title: string | ReactNode,
     keypadClickHandlers: KeypadClickHandlers,
-    passcode: string,
+    passcode: PasscodeInputProps,
     nextButtonAction?: () => void,
     previousButtonAction?: () => void
 } & PageTransitionProps
 
-const PasscodePage = ({ title, keypadClickHandlers, passcode, nextButtonAction, previousButtonAction, pageTransition}: PasscodeProps) => {
+const PasscodePage = ({
+                          title,
+                          keypadClickHandlers,
+                          passcode,
+                          nextButtonAction,
+                          previousButtonAction,
+                          pageTransition
+                      }: PasscodeProps) => {
     return (
         <Page
             title={title}
             contentComponents={
-              <>
-              <PasscodeInput value={passcode}/>
-                  <Keypad clickActionHandlers={keypadClickHandlers}/>
-              </>
+                <>
+                    <PasscodeInput {...passcode}/>
+                    <Keypad clickActionHandlers={keypadClickHandlers}/>
+                </>
             }
-            buttonComponents={<StepButtons nextButton={{action: nextButtonAction}} previousButton={{action: previousButtonAction}}/>}
+            buttonComponents={<StepButtons nextButton={{action: nextButtonAction}}
+                                           previousButton={{action: previousButtonAction}}/>}
             pageTransition={pageTransition}
         />
     )