+++ src/components/UI/Inputs/PasscodeInput/PasscodeInput.tsx
@@ -2,10 +2,11 @@ import styled from 'styled-components';
 import {RiErrorWarningFill} from "react-icons/all";
 import {FormattedMessage} from "react-intl";
 
-type PasscodeInputProps = {
+export type PasscodeInputProps = {
     value: string,
     error?: boolean,
-    invalid?: boolean
+    confirmError?: boolean,
+    lengthError?: boolean
 }
 
 const StyledPasscode = styled.div`
@@ -24,9 +25,11 @@ const StyledInput = styled.input`
   padding: 10px;
   width: 16rem;
   text-align: center;
+
   &:focus {
     outline: none;
   }
+
   &:disabled {
     border-bottom: 1px solid ${props => props.theme.ui.passcodeInput.border};
     color: ${props => props.theme.ui.passcodeInput.color};
@@ -40,20 +43,24 @@ type StyledErrorMessageProps = {
 
 const StyledErrorMessage = styled.p<StyledErrorMessageProps>`
   color: ${props => props.theme.ui.passcodeInput.color};
-  font-size: 14px;
+  font-size: 16px;
   display: flex;
   align-items: center;
   transition: opacity 100ms ease-in-out;
   opacity: ${props => props.show ? 1 : 0};
+  margin-top: 10px;
 `
 
 const PasscodeInput = (props: PasscodeInputProps) => {
     return (
         <StyledPasscode>
             <StyledInput type="password" value={props.value} disabled={true}/>
-            <StyledErrorMessage show={props.error || props.invalid}><RiErrorWarningFill style={{marginRight: '5px'}} size={16}/>
-                {props.invalid ? <FormattedMessage id="unlock.invalid"/> : null}
-                {props.error ? <FormattedMessage id="unlock.error"/> : null}
+            <StyledErrorMessage show={props.error || props.confirmError || props.lengthError}><RiErrorWarningFill
+                style={{marginRight: '5px'}}
+                size={16}/>
+                {props.lengthError ? <FormattedMessage id="passcodeInput.invalidLength"/> : null}
+                {props.confirmError ? <FormattedMessage id="passcodeInput.confirm.error"/> : null}
+                {props.error ? <FormattedMessage id="passcodeInput.incorrect.error"/> : null}
             </StyledErrorMessage>
         </StyledPasscode>
     )