+++ src/components/UI/Layout/Pages/Locked-Register/Page.tsx
@@ -1,5 +1,5 @@
-import { motion } from 'framer-motion'
-import { ReactNode } from 'react'
+import {motion} from 'framer-motion'
+import {ReactNode} from 'react'
 import styled from 'styled-components'
 import {pageTransitionVariants} from "../../../Motion/Variants/Variants";
 
@@ -27,7 +27,6 @@ const StyledAnimatedContainer = styled(motion.div)`
   flex-direction: column;
   justify-content: space-between;
   align-items: center;
-  padding: 50px 0;
   background-color: ${props => props.theme.ui.backgroundColor};
 `
 
@@ -36,15 +35,20 @@ const StyledAnimatedSection = styled(motion.div)<StyledSectionProps>`
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  min-width: 20rem;
+  width: 100%;
+  max-width: 60rem;
   flex: ${props => props.flex || 1};
-  padding: 10px 0;
+  overflow-y: scroll;
 `
 
 const StyledTitle = styled.h1`
-    
+  background-color: ${props => props.theme.ui.backgroundColor};
+  width: 100%;
+  padding: 30px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 `
-
 
 
 const Page = ({title, contentComponents, buttonComponents, pageTransition}: PageProps) => {
@@ -57,17 +61,19 @@ const Page = ({title, contentComponents, buttonComponents, pageTransition}: Page
             exit="exit"
             variants={pageTransitionVariants}
             transition={{
-                x: { type: "just" },
-                opacity: { duration: 0.2 }
+                x: {type: 'just'},
+                opacity: {duration: 0.2}
             }}
         >
-            <StyledTitle>
-                {title}
-            </StyledTitle>
+            {title && (
+                <StyledTitle>
+                    {title}
+                </StyledTitle>
+            )}
             <StyledAnimatedSection flex={1}>
                 {contentComponents}
             </StyledAnimatedSection>
-            <StyledAnimatedSection flex={0.05}>
+            <StyledAnimatedSection flex={0.05} style={{padding: '30px 10px'}}>
                 {buttonComponents}
             </StyledAnimatedSection>
         </StyledAnimatedContainer>