+++ src/localization/messages/en-US.ts
@@ -1,15 +1,16 @@
-import { Messages } from '../types'
+import {Messages} from '../types'
 
 const en_US: Messages = {
     'main.greeting': 'Hello world!',
     'onboarding.selectLanguage': 'Select Language',
     'onboarding.setPasscode': 'Create Passcode',
     'onboarding.confirmPasscode': 'Confirm Passcode',
+    'passcodeInput.incorrect.error': 'Incorrect passcode, please try again!',
+    'passcodeInput.confirm.error': 'Passcode not correct, please try again!',
+    'passcodeInput.invalidLength': 'Passcode requires minimum 6 numbers',
     'button.next': 'Next',
     'button.back': 'Back',
     'unlock.title': 'Enter your Seguro passcode',
-    'unlock.error': 'Incorrect passcode, please try again!',
-    'unlock.invalid': 'PasscodePage requires 6 numbers',
     'keypad.cancel': 'Cancel',
     'keypad.unlock': 'Unlock'
 }