+++ src/localization/messages/zh-CN.ts
@@ -1,15 +1,16 @@
-import { Messages } from '../types'
+import {Messages} from '../types'
 
 const zh_CN: Messages = {
     'main.greeting': '你好世界!',
     'onboarding.selectLanguage': '选择语言',
     'onboarding.setPasscode': '设置密码',
     'onboarding.confirmPasscode': '确认密码',
+    'passcodeInput.incorrect.error': '密码错误，请再试一次',
+    'passcodeInput.confirm.error': '密码不对，请再输入一次',
+    'passcodeInput.invalidLength': '密码最少需要6位数',
     'button.next': '下一步',
     'button.back': '退一步',
     'unlock.title': '输入您的 Seguro 密码',
-    'unlock.error': '密码错误，请再试一次',
-    'unlock.invalid': '密码最少需要6位数',
     'keypad.cancel': '取消',
     'keypad.unlock': '解锁'
 }